+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.0.6
+- Cosmetic changes
+- [Added] `checkboxFillColor` checkbox fill color
+- [Added] `splashColor` splash color on list tile
+- [Added] `listTextStyle` text style for text on list tile
+
 # 0.0.5
 - Optimization
 - [Added] All parameters in readme.
@@ -5,11 +11,11 @@
 # 0.0.4
 - [Added] `duration` parameter for debounce in search option
 
-## 0.0.3
+# 0.0.3
 - [Added] `textStyle` parameter for the anchor element
 
-## 0.0.2
+# 0.0.2
 - Updated README.md file
 
-## 0.0.1
+# 0.0.1
 - Initial development release.